+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning][].
 ## [0.2.2]
 
 ### Feature
-- Only `X`, `obs`, and `var` are currently yielded when reading a {class}`~annbatch.DatasetCollection` or adding data via {meth}`~annbatch.Loader.add_adata`/{meth}`~annbatch.Loader.add_adatas`. When observation-aligned {attr}`~anndata.AnnData.obsm`, {attr}`~anndata.AnnData.obsp`, or {attr}`~anndata.AnnData.layers` elements are present, they are dropped for now and a {class}`FutureWarning` is emitted; a **future release will yield them as well**. To silence the warning, drop these elements beforehand (e.g. via a custom `load_adata` for {meth}`~annbatch.Loader.use_collection`).
+- Only `X`, `obs`, and `var` are currently yielded when reading a {class}`~annbatch.DatasetCollection` or adding data via {meth}`~annbatch.Loader.add_adata`/{meth}`~annbatch.Loader.add_adatas`. When observation-aligned {attr}`~anndata.AnnData.obsm`, {attr}`~anndata.AnnData.obsp`, or {attr}`~anndata.AnnData.layers` elements are present, they are not yielded for now and a {class}`FutureWarning` is emitted; a **future release will yield them as well**. To silence the warning, drop these elements beforehand (e.g. via a custom `load_adata` for {meth}`~annbatch.Loader.use_collection`).
 
 ## [0.2.1]